Macro calypso_diagnostic::gen_error

macro_rules! gen_error {
    (Err($($rest:tt)*) as $ty:ty) => { ... };
    (sync $grcx:expr, $($rest:tt)*) => { ... };
    (@i3 $diagnostic:ident; notes: [$($note:expr),*$(,)?]) => { ... };
    (@i2 $diagnostic:ident; labels: [$($style:expr => ($source:expr, $span:expr); $message:expr),*$(,)?]$(, $($rest:tt)*)?) => { ... };
    (@fmt $diag:expr, $($name:ident = $value:expr),*) => { ... };
    (@fmt $diag:expr) => { ... };
    (@i1 $self:expr => { $code:ident$(, $($name:ident = $value:expr),*)?; $($rest:tt)* }) => { ... };
}

Generate errors or report synchronized errors.

Panic errors are handled with Results. They can be generated as follows:

gen_error!(Err(self => {
    E0005, fmt_values = "values";
    labels: [
        LabelStyle::Primary =>
            (source_id, span);
            "label message",
        LabelStyle::Secondary =>
            (source_id2, span2);
            "label message 2"
    ],
    notes: [
        "note 1",
        "note 2"
    ]
}) as ())?

The as () at the end is to change the Ok type of the result, if necessary. Panic errors should only be used if there is ABSOLUTE CONFIDENCE that there is NO way to recover from this error.